[FIX] fleet_vehicle_log_fuel: Don't replace whole SQL view + make it work + renaming
Replacing whole SQL view definition is totally incompatible with other
modules adding things to this report. We use a trick for replacing
only some parts retrieving the current SQL view for keeping the
inheritability.

We need as well to add the new selection value.

Finally, a renaming of the model file for proper recognition has been
done.

TT48062